separate tests for models and controllers based on app
